requirements.txt
ethereum>=2.3.2
py-solc>=3.2.0
requests>=2.19.1
web3>=4.8.2